refactor: rename modules to polymod_{crc32,rle} to avoid header clashes
Generic header names crc32.h and rle.h in the QMK include search path
were too easy to collide with other modules or future code. Prefix the
module dir, .c, .h, and all #include sites with polymod_ to namespace
them.

Underscore (not hyphen) is required because QMK's community-modules
generator uppercases the directory name into C macros like
COMMUNITY_MODULE_<NAME>_ENABLE and SPLIT_TRANSACTION_IDS_MODULE_<NAME>,
which must be valid C identifiers. All upstream QMK modules use
underscores (hello_world, super_alt_tab, etc.).

Verified clean builds on both split72 and corne42.
