+++ keyboards/handwired/polykybd/corne42/keyboard.json
@@ -11,8 +11,8 @@
         "device_version": "1.0.0"
     },
     "modules": [
-        "polykybd/crc32",
-        "polykybd/rle"
+        "polykybd/polymod_crc32",
+        "polykybd/polymod_rle"
     ],
     "features": {
         "bootmagic": false,
