+++ keyboards/handwired/polykybd/split_sync.c
@@ -13,7 +13,7 @@
 #include "base/com.h"
 #include "base/disp_array.h"
 #include "base/update.h"
-#include "crc32.h"
+#include "polymod_crc32.h"
 #include "fill_overlay.h"
 #include "state.h"
 #include "side.h"
