+++ modules/polykybd/polymod_crc32/polymod_crc32.h
