+++ modules/polykybd/polymod_crc32/qmk_module.json
